update feeding amount to double

+++ src/main/kotlin/com/babytrackr/analytics/controller/model/FeedData.kt
@@ -2,5 +2,5 @@ package com.babytrackr.analytics.controller.model
 
 data class FeedData(
     val totalFeedings: Int,
-    val totalOunces: Int
+    val totalOunces: Double
 )
